fixed bug
There was a bug, where the lamp hours would "ER402" while the projector was powered on (but would show up just fine while it was powered off). Apparently, the PT-FW430 doesn't like my specification of "Lamp 1" (it's the only lamp) while the projector is on, but doesn't mind my doing so when the projector is off.